backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic[email]>=2.6.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
pydfs-lineup-optimizer>=3.6.0
aiofiles>=23.2.1
# Auth dependencies (added v1)
# Note: use bcrypt directly — passlib 1.7.4 is incompatible with bcrypt 4.x+
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
stripe>=8.0.0
slowapi>=0.1.9
python-dotenv>=1.0.0
# PostgreSQL
psycopg2-binary>=2.9.9
# Migrations
alembic>=1.13.0
